Enabled LSM9DS1 sensor and forced module name
To maintain a single build target that is compatible with all of
the ventana products, the module name was forced to sensors.stm as to
not conflict with a preexisting sensors.ventana module.

Enabled the LSM9DS1 magnetometer/accelerometer/gyroscope which is
currently used on the latest revisions of the GW5530, GW5600, and
GW5904.

Signed-off-by: Robert Jones <rjones@gateworks.com>